+++ doc/source/changelog.md
@@ -23,6 +23,9 @@ Semver is not yet in place, so each version can have breaking changes, although 
         - Fixed slow loading when the database path contains `res://` or `user://`
     - `VoxelInstancer`: Fixed instances with LOD > 0 were generated on `VoxelTerrain` even though LOD isn't supported (ending up in weird positions). No instances should generate.
 
+- Breaking changes
+    - `VoxelVoxLoader`: methods are now static, so no instance of the class need to be created
+
 
 1.2 - 20/04/2024 - branch `1.2` - tag `v1.2.0`
 ------------------------------------------------